% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dnase_data}
\alias{dnase_data}
\title{DnaseI hypersensitivity data}
\format{A list of 20 dataframes with about 150,000-200,000 rows and 4 columns each:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{name}{DnaseI hypersensitivity score}
}}
\source{
\url{https://s3.amazonaws.com/bedtools-tutorials/web/maurano.dnaseI.tgz}
}
\usage{
dnase_data
}
\description{
A dataset containing DNaseI hypersensitivity data from 20 human cell types
}
\references{
Maurano et al. Systematic Localization of Common Disease-Associated Variation in Regulatory DNA. Science. 2012. Vol. 337 no. 6099 pp. 1190-1195.
}
\seealso{
\url{http://www.sciencemag.org/content/337/6099/1190.short}
}
\keyword{datasets}

