% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hg19_genes}
\alias{hg19_genes}
\title{Human RefSeq annotations}
\format{A data frame with 24,062 rows and 6 variables:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{name}{gene symbol}
  \item{score}{0}
  \item{strand}{strand}
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables}
}
\usage{
hg19_genes
}
\description{
A dataset containing RefSeq gene annotations for hg19
}
\keyword{datasets}
