% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{repeat_data}
\alias{repeat_data}
\title{Human repetitive elements}
\format{A data frame with 5,520,017 rows and 8 variables:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{name}{repeat name}
  \item{score}{Smith-Waterman alignment score, i.e. swScore}
  \item{strand}{strand}
  \item{repClass}{repeat class}
  \item{repFamily}{repeat family}
}}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables}
}
\usage{
repeat_data
}
\description{
A dataset containing annotations for repeat classes in the hg19 human genome, as
annotated by RepeatMasker
}
\keyword{datasets}

