% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{segway_data}
\alias{segway_data}
\title{Segway encyclopedia of human regulatory elements}
\format{A data frame with 164,785 rows and 169 columns
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{sum_score}{sum of functionality score}
  \item{mean_score}{average base-wise functionality score}
  \item{...}{segway label of each cell type in the interval}
}}
\source{
\url{http://noble.gs.washington.edu/proj/encyclopedia/segway_encyclopedia.bed.gz}
}
\usage{
segway_data
}
\description{
A dataset containing segway annotations for 164 human cell types
}
\references{
Libbrecht MW, Rodriguez O, Hoffman MM, Bilmes JA, Noble WS. 2016. A unified encyclopedia of human functional elements through fully automated annotation of 164 human cell types. Biorxiv preprint: http://dx.doi.org/10.1101/086025.
}
\seealso{
\url{http://noble.gs.washington.edu/proj/encyclopedia/}
}
\keyword{datasets}
