% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tutorial_files}
\alias{tutorial_files}
\title{BEDTools Tutorial files}
\format{A list of dataframes
\describe{
  \item{cpg}{dataframe with bed intervals of CPG islands in human genome}
  \item{exons}{dataframe with bed intervals for exons in human genome}
  \item{gwas}{dataframe of  bed intervals of disease associated snps in human genome}
  \item{hesc_chromHmm}{dataframe of bed intervals with predicted functions for intervals in human genome}
  \item{genome}{dataframe of chromosome length for human genome}
}}
\source{
\url{http://quinlanlab.org/tutorials/bedtools/bedtools.html}
}
\usage{
tutorial_files
}
\description{
A list of BED and genome files used by the BEDTools tutorial
}
\keyword{datasets}
